Rename to ReciprocalLatticePoint, other minor changes
